new features and bugfixes
Changes:
Removed the original 'listrules.py' from this branch.
Renamed 'newlistrules.py' to 'listrules.py'.
Changed csv 'csvheader' option to 'headeroff' and set the default to including a header to ensure the same default behaviour as listrules.py v1.0 was persisted.
Changed VALID_PERMISSIONS from being a set to a list to ensure the permissions order returned is consistent.
Renamed columns and results for:
createdTimestamp --> creationTimeStamp
modifiedTimestamp --> modifiedTimeStamp

New features:
Option to print filter and REST call without submitting
Option to filter by createdBy and modifiedBy

Bugfix:
Fixed a bug (through the column and result renamed) affecting the created* and modified* timestamp filters didn't work.
